When launching debug configuration, prepend CWD to program path.
For some reason, when the 'program' specified is a relative path, the (cppdbg) debugger is not able to
find the program, eventhough the path is correct and the current working directory is correct.
So this is a work-around: Prepend current working directory to 'program' - if 'program' is a relative path.